Lab02 Update++

+++ WebBanHang/Controllers/ProductController.cs
@@ -147,12 +147,24 @@ namespace WebBanHang.Controllers
             ViewBag.TL = _db.Categories.Select(x => new SelectListItem { Value = x.Id.ToString(), Text = x.Name });
             return View();
         }
-
-        //delete
         public IActionResult Delete(int id)
+        {
+           var sp = _db.Products.Find(id);
+            if (sp == null)
+            {
+                return NotFound();
+            }
+            return View(sp);
+        }
+        //delete
+        public IActionResult DeleteConfirmed(int id)
         {
             //doc thong tin san pham can xoa
             var sp = _db.Products.Find(id);
+            if (sp==null)
+            {
+                return NotFound();
+            }
             //xoa khoi csdl
             _db.Products.Remove(sp);
             _db.SaveChanges();
@@ -165,7 +177,7 @@ namespace WebBanHang.Controllers
                     System.IO.File.Delete(oldFilePath);
                 }
             }            //thong bao thao tac co the thanh cong hoac that bai
-            TempData["success"] = "Deleted";
+            TempData["error"] = "Deleted";
             //dieu huong ve action Index
             return RedirectToAction("Index");
